requirements.txt
python-dotenv
mysql-connector-python
RPi.GPIO